транзакция

+++ internal/models/seat.go
@@ -15,7 +15,7 @@ type Place struct {
 	Name         string    `json:"name" gorm:"not null" binding:"required,min=2"`
 	Type         PlaceType `json:"type" gorm:"not null" binding:"required,oneof=workspace meeting_room"`
 	Description  string    `json:"description"`
-	PricePerHour float64   `json:"price_per_hour" gorm:"not null" binding:"required,gt=0"`
+	PricePerHour int       `json:"price_per_hour" gorm:"not null" binding:"required,gt=0"` // в копейках
 	IsActive     bool      `json:"is_active" gorm:"not null;default:true"`
 	CreatedAt    time.Time `json:"created_at"`
